Have record() initialize itself, like replay() does.
